--- src/bootx64/src/memory.rs
@@ -1,2 +1,0 @@
-pub mod map;
-
